Rename Microsoft.Build.Tasks.SourceControl to Microsoft.SourceLink.Common and fix dependencies.
