+++ src/Microsoft.Build.Tasks.Git.UnitTests/GitOperationsTests.cs
@@ -5,9 +5,9 @@ using System.IO;
 using System.Linq;
 using LibGit2Sharp;
 using Microsoft.Build.Framework;
-using Microsoft.Build.Tasks.SourceControl.UnitTests;
+using Microsoft.SourceLink.Common.UnitTests;
 using Xunit;
-using static Microsoft.Build.Tasks.SourceControl.UnitTests.KeyValuePairUtils;
+using static Microsoft.SourceLink.Common.UnitTests.KeyValuePairUtils;
 
 namespace Microsoft.Build.Tasks.Git.UnitTests
 {
