+++ src/SourceLink.Common.UnitTests/PathUtilitiesTests.cs
@@ -1,8 +1,9 @@
 ﻿// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
 using Microsoft.Build.Tasks.Git.UnitTests;
+using Microsoft.Build.Tasks.SourceControl;
 using Xunit;
 
-namespace Microsoft.Build.Tasks.SourceControl.UnitTests
+namespace Microsoft.SourceLink.Common.UnitTests
 {
     public class PathUtilitiesTests
     {
