+++ src/SourceLink.Common/xlf/Resources.de.xlf
