+++ src/SourceLink.Common/xlf/Resources.es.xlf
