+++ src/SourceLink.Common/xlf/Resources.fr.xlf
