+++ src/SourceLink.Common/xlf/Resources.pt-BR.xlf
