+++ src/SourceLink.GitHub/GetSourceLinkUrl.cs
@@ -2,7 +2,6 @@
 
 using System;
 using System.Collections.Generic;
-using System.IO;
 using System.Linq;
 using Microsoft.Build.Framework;
 using Microsoft.Build.Tasks.SourceControl;
