+++ src/SourceLink.GitHub/Microsoft.SourceLink.GitHub.nuspec
@@ -14,7 +14,7 @@
     <serviceable>$Serviceable$</serviceable>
     <repository type="$RepositoryType$" url="$RepositoryUrl$" commit="$RepositoryCommit$" />
     <dependencies>
-      <dependency id="Microsoft.Build.Tasks.SourceControl" version="$Version$" />
+      <dependency id="Microsoft.SourceLink.Common" version="$Version$" />
       <dependency id="Microsoft.Build.Tasks.Git" version="$Version$" />
     </dependencies>
   </metadata>
