+++ src/SourceLink.Vsts.Git/Microsoft.SourceLink.Vsts.Git.nuspec
@@ -14,6 +14,7 @@
     <serviceable>$Serviceable$</serviceable>
     <repository type="$RepositoryType$" url="$RepositoryUrl$" commit="$RepositoryCommit$" />
     <dependencies>
+      <dependency id="Microsoft.SourceLink.Common" version="$Version$" />
       <dependency id="Microsoft.Build.Tasks.Git" version="$Version$" />
     </dependencies>
   </metadata>
